new feature: set variables temporarily using `with` statement
The new `__call__()` method implements a context manager
protocol. Basically it allows to do this:

    with current(locale='en_US'):
        do_something_useful(current.locale)

Upon exit, modified attributes are restored to their
previous values using the `ContextVar.rest()` method.

Also, I re-named and slightly re-organized methods in the class
to make them more readable in the code, and also look better
in the auto-generated documentation.