replace Token.MISSING with my custom Missing object
The Token.MISSING object was designed for use with Token objects,
and using it elsewhere looks odd. Also, there are no guarantees
that its implementation won't change in the future.

Initially, I used Token.MISSING for my own purposes because I wanted
to get started quickly.

Now, after a while, I decided to do a little refactoring and replace
it with my own `Missing` object.

Maybe it doesn't make a difference technically,
but this way it just looks cleaner to me.