move error texts to docstrings of exception classes
Previously, I just raised an exception, passing text as argument.

That worked, but auto-generated documentation was ugly,
like a bunch of exception sub-classes with no descriptions at all.

I decided to add docstrings, and since I'm a bit lazy,
I decided to re-use the exception class docstring
as the error message supplied when throwing exception in Python.

So, there is the new class:

    ExceptionDocstringMixin

it automatically uses class docstring as an error message.

And then I added the mixin to all exceptions, improved the texts,
and now they look good in auto-generated docs, and also look decent
when thrown in Python.