fix hasattr()/getattr() exception when variable is not set
Problem: an attempt to read an unitialized variable,
it raises LookupError.

The LookupError is raised not only when you just read the attribute,
but also when you call `getattr()` or `hasattr()`, which is
a bit odd.

My expectation is that `getattr()`/`hasattr()` should not
raise an exception if the variable is not set.

So to make the code less surprising, I'm now catching
`LookupError` and re-throwing it as `AttributeError`.

...with one extra trick: not really an `AttributeError`,
but a subclas, based on *both* `LookupError` and `AttributeError`.
That allows to fix `getattr()`/`hasattr()`, and in the same time
preserve consistency with the `contextvars` module.