move docstring from class to module
No changes in the text, just move the docstring to the module.

I have to do this because the docstring bloats, and I'm going to split
it to sub-sections, and Sphinx doesn't support sub-sections in class
docstrings.

So, in order to have sections in the docstring, I have to move the
text form the class to the module docstring.