add support for deleting values of context variables
Python doesn't provide any standard way to delete or erase
context variables, so I had to implement it in a hacky way:

When you call `.delete()` (or `__del__`), nothing is actually deleted.
Instead, a special `ContextVarValueDeleted` marker object is written
to the context variable.

Later on, when you call `.get()`, it detects the marker,
and behaves as if there was no value at all.

So now deletes are supported, although they're a bit slow.

I tried to reduce performance hit of a `.get()` call
as much as possible, but it still much slower than the
built-in (and written in C) ContextVar.get method.

So, in case somebody needs performance, I've left
the `ContextVarDescriptor.get_raw` method, which is a direct
reference to the lower-level C method.