remove doctest-style `>>>` prompts from README.rst
...because I have to remove them manually when I copy-paste
examples from README to Python shell.

Well, yeah, that doesn't work with `doctest`, that is,
without `>>>`, code is not executed as a test,
but I'm fine with it, because it is an issue only in README.rst,
that is displayed on GitHub repo page.

For other .rst files it is not a problem, because they're
part of Sphinx-generated docs (and not a GitHub page),
and in Sphinx, I use the `sphinx-togglerompt` extension
that allows to hide these nasty `>>>` things.

So, examples in `README.rst` will not be executed by `doctest`,
but I can deal with it.