get rid of custom ParamsDict in favor of inspect.Signature.parameters
Problem: the code uses 2 different ways to describe parameters:

 1. inspect.Signature.parameters
 2. ParamsDict (a custom type that was added by me)

It is confusing to have 2 different representations at the same time.

In the past, I thought it is a good idea to have a custom
`ParamsDict`, because I could extend it with more fields,
like the `position` (which is not available in `inspect.Parameter`).

Now I think that the custom `ParamsDict` confuses more than helps,
so I decided to get rid of it.

The drawback is that the `_generate_rules_for_single_source()`
function became more complex.

But the net result is that the module became simpler overall,
because that confusing `ParamsDict` is removed from the module.