add python_contextvars.rst - extra docs for the standard contextvars
There are some non-ovious things about the standard ContextVar
objects that should have been mentioned in the official docs,
but for some reason they haven't.

So, I decided to add a page, where I'm having extra docs for the
standard contextvars. Various caveats, misconceptions and recipes,
that are missing in the official docs.