rename: args_from_context -> supply_args
...because the decorator is designed to be decoupled
from context variables.

Context variables is only one of possible sources,
sort of an implementation detail, so the name shouldn't
depend on it.

Furthermore, I'm thinking about extracting this decorator
to a separate Python package (that doesn't depend on context
variables), so the word "context" would look odd there.

So, I've chosen the new name:

    @supply_args