rename InjectionRuleTuple -> SupplyRuleTuple
...also rename all related function/variable names,
docstrings, and all coments where "inject" word occurs.

The word "inject" comes from old times, when the
`@supply_args` decorator was called `@inject_args`.

I don't want to use the word "inject" (because I don't want to
interfere with Dependency Injection terminology and thus confuse
programmers who could read the code).

...so I decided to walk across the code and clean up occurrences
of the word "inject", re-naming it to "supply" or something else.

So, there is a bunch of re-names in this commit.