extract `@supply_args` decorator to a separate Python package
Previously it was the `args.py` file in this `contextvars-extras`
git repo, but then I decided to extract it to a separate Python
package.

So, the `args.py` is removed, and all its functions are now
imported from the `supply_args` (an external Python pakcage,
where I moved the code already).