drop support of `@registry.as_args` and `@supply_args` decorators
These decorators allow to implicitly pass registry as arguments
to functions.

Previously I thought it is a good idea, but I changed my mind.

Now I think that it provokes writing bad code, where, for example,
adding a new variable to registry may affect functions,
that (by an accident) have arguments with the same name.

I could improve and re-write it in a more explicit manner,
but I don't want to focus on that now.

For now, I decided to just drop support for the `@supply_args`.
Maybe later I will introduce it again in a different form,
or maybe not :/