improve documentation for contextvars_extras.descriptor module
Previously it was just a bunch of docstrings, which was
hard to read and navigate.

So, I moved these docstrings to a separate .rst file,
and added a lot of new text there. So now there is a full-weighted
page, located in a seaparate .rst file, which is much easier
to navigate.