split ContextVarDescriptor and ContextVarExt classes
Previously there was 1 class:
 - ContextVarDescriptor (located in `descriptor.py`)

and here I'm splitting it into 2 pieces:
 - ContextVarExt (`context_var_ext.py`)
 - ContextVarDescriptor (`context_var_descriptor.py`)

The majority of code went to `ContextVarExt` (the base class),
and the descriptor-related logic is placed in
`ContextVarDescriptor` (the subclass).

So now, I magine, it became easier to understand these classes:
you first read the documentation for the `ContextVarExt`
(and be like "ok, this is just an extended ContextVar object),
and then you read the next page for the `ContextVarDescriptor`
(and be like: "ok, this is just `ContextVarExt` that also
behaves like a `@property`).

But, this is not the main reason why I've done it.
The main reason is that `__init__()` methods are different
in these two cases (descriptor vs standalone object).

In case of descriptor, you need to postpone initialization
of ContextVar until the `__set_name__` method is called
(whereas in standalone object, you don't need to postpone it).

That could have been solved by passing a flag to `__init__()`,
or some alternative constructor methods, but that looked ugly to me,
while a separate subclass looks nice and simple.

So, I decided to split two classes, and soon that will allow
to have two different `__init__()` implementations.