new function: get_context_var_default()
For some reason, the built-in ContextVar class doesn't
allow to reach its default value.

So, I had to write a function for extracting the
default value in a tricky way.

I need it for some ContextVarExt features,
but it also is a useful function that can be
useful in the outside world.