rename file: `context.py` to `context_management.py`
The word "context" is too ambiguous, because everything in this
package is somehow related to context variables, so this single
word doesn't tell you much.

The new name "context_management.py` is also a bit ambiguous
(can be confused with Python's context managers concept),
but this is how it is called in the Python's `contextvars` package,
(there is the "Manual Context Management" section in the docs),
so I'm using the same terminology here.