get rid of `misc.py`, moving its code to `context_var_ext.py`
Because `misc` is too general name that doesn't tell you anything,
and this module is just doomed to bloat with junk in the future.

So, I decided to prevent it early, moving the single function that
lives there to `context_var_ext.py` - the module that relates the most
to the function.