doc: remove docstring that looks very odd in Sphinx docs
It looks odd because module's documentation page
has a non-standard layout, and so this docstring appears
in a place that is not related to the `ContextVarExt` class,
and that looks terrible.

The `sphinx.ext.autodoc` doesn't provide an easy way to hide it,
so I have to delete the docstring.