fix: reset_to_default() doesn't when re-using existing ContextVar
Problem: when you pass an existing ContextVar() object
to ContextVarExt() wrapper, the `.reset_to_default()` method
doesn't work.

This is just because the standard `contextvars` module doesn't
provide any out-of-the-box API to extract a default value
from an existing `ContextVar` object.

But, recently, I've added the `get_context_var_default()` utility
function, that extracts the default value (in a tricky way),
so here I'm using this fuction to fix the issue with
`.reset_to_default()`.