fix: inconsistent defaults with existing ContextVar object
Problem: there is a possibility of initializing ContextVarExt()
in such way that its default value will be different form
the default value of the underlying ContextVar object.

When that happens, the ContextVarExt() object starts to behave
inconsistently, because in some cases ContextVarExt._default
is used, while in other cases ContextVar's internal default
value is used.

To avoid such bugs, I have to a couple of restrictions:

1. these two arguments cannot be passed together:
   `ContextVarExt(context_var, default)`
   (now `default` is always derived from `context_var`)

2. `deferred_default` cannot be passed if the underlying
   ContextVar has a default value.