refactor: allow to to use generic types without arguments
That is, instead of this:

    ContextVar[int]

I can now just write this:

    ContextVar

This is done because on this project I have a lot of containers
(`ContextVarExt`, `ContextVarDescriptor`, `ContextVarsRegistry`)
that can store anything inside them.

That is, in most cases I have to write `ContextVarsRegistry[Any]`,
and this `Any` is just useless noise, that makes the code ugly,
and doesn't bring any benefit.

So, I decided to make this `[Any]` part optional.