refactor: remove `Generic[VarValueT]` from `ContextVarsRegistry`
Because for me, it is hard to imagine a real-world use case,
where all variables in `ContextVarsRegistry` will have the same type.

That is, you almost always have to write:

    class MyRegistry(ContextVarsRegistry[Any]):
        ...

So this `[Any]` is just useless noise that doesn't bring any benefit,
and therefore I decided to drop this parameter entirely.

Now `ContextVarsRegistry` is not parametrized at all.