refactor: remove useless type defs from internal_utils.py
These type definitions are not used anymore.

They were, but it turned out that mypy doesn't understand
such aliases for generic types. So I had to walk over all decorators
on the project and replace them with ad-hoc `Callable[...]`
annotations.

So, long story short: now this is dead code, and I can delete it.