refactor: replace `MISSING` with `NO_DEFAULT` in context_var_ext.py
Problem: it is tempting to use `MISSING` as the argument
to the `.get()` method, like this:

    value = context_var.get(MISSING)
    if value is MISSING:
        ...

But, that leads to a confusing bug, because the default value for the
parameter of `.get()` is already `MISSING`, and it already indicates
"no default value" condition, which triggers `LookupError`.

So the bug is that: `LookupError` is raised even though
the `.get()` method shall not raise it if you pass a default value.

And a similar bug occurs in other places, where ContextVar's default
appears (like for example the `ContextVarExt.__init__()` method).

So, to fix that, I'm inroducing a new sentinel object:

    contextvars_extras.context_var_ext.NO_DEFAULT

and now it is used in all places related to default values.

So now it all looks much less confusing,
and you can use `MISSING` with the `.get()` method.