fix: ContextVarsRegistry.keys() triggers deferred_default
The `.keys()` must be a read-only method, but right now it is a bit
read-write because of the magical `deferred_default`.

Here I'm changing it slightly, such that `.keys()` doesn't trigger the
`deferred_default`, and thus never alters state of the registry.

The code is a bit copy, but it works.